requirements.txt
aioredis==1.3.1
asgiref==3.3.2
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
celery==5.2.3
coreapi==2.3.3
coreschema==0.0.4
constantly==15.1.0
cryptography==36.0.0
defusedxml==0.7.0rc1
daphne==3.0.2
Django==3.2.4
djangochannelsrestframework==0.3.0
django-cors-headers==3.6.0
django-celery-beat==2.5.0
django-celery-results==2.0.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
django-oauth-toolkit==2.2.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
google-auth==2.17.0
h11==0.12.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hypercorn==0.12.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.3
oauthlib==3.1.0
Pillow==8.0.1
priority==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pyOpenSSL==21.0.0
pytz==2021.3
redis==4.5.1
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==1.4.1
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0

